Subo ultimo commit para mostrar v31 MercadoPago

+++ src/app/components/shared/side-cart/side-cart.component.ts
@@ -22,34 +22,54 @@ export class SideCartComponent implements OnInit, DoCheck {
     private productoService: ProductosService,
     private carritoService: CarritoService
   ) {
+    this.productosCarrito["total"] = 0;
     if (localStorage.getItem('userId')) {
-      const productosCarrito = this.carritoService.getCarrito();
-        console.log(productosCarrito);
-      this.productosCarrito["total"] = 0;
+      // const productosCarrito = this.carritoService.getCarrito();
+      let carritoDeComprasJsonUsuario = [];
+      setTimeout(() => {
+        if (localStorage.getItem("carritoDeCompras")) {
+          const carritoDeComprasJson = JSON.parse(
+            localStorage.getItem("carritoDeCompras")
+          );
 
-      if (productosCarrito !== null) {
-        
-        productosCarrito.forEach((productoCarrito: any) => {
-          const todosLosProductosJson = JSON.parse(localStorage.getItem('todosLosProductos'));
-
-          todosLosProductosJson.forEach(producto => {
-            if (producto.id == productoCarrito.productId) {
-              let pathImagenDetalle: any[] = [];
-              pathImagenDetalle = this.productoService.imagenesDetalle(producto.id);
-              // this.productoService.getImagenesDetalle(producto.id).subscribe((res: any) => {
-              //     res.forEach(imagen => {
-                    // pathImagenDetalle.push(imagen.path);
-                  // });
-                // });
-              producto.idCarrito = productoCarrito.id;
-              producto.path = pathImagenDetalle;
-              producto.talle = productoCarrito.talle;
-              // this.productosCarrito['total'] = this.productosCarrito['total'] + producto.precio;
-              this.productosCarrito.push(producto);
+          carritoDeComprasJson.forEach(carrito => {
+            if (
+              carrito.userId == localStorage.getItem("userId") &&
+              carrito.orden_id === 0
+            ) {
+              carritoDeComprasJsonUsuario.push(carrito);
             }
           });
-        });
-      }
+        }
+      }, 1500);
+      console.log(carritoDeComprasJsonUsuario);
+      setTimeout(() => {
+        if (carritoDeComprasJsonUsuario !== null) {
+          carritoDeComprasJsonUsuario.forEach((productoCarrito: any) => {
+            console.log("entra1");
+            if (productoCarrito.orden_id === 0) {
+              console.log("entra2");
+              const todosLosProductosJson = JSON.parse(
+                localStorage.getItem("todosLosProductos")
+              );
+
+              todosLosProductosJson.forEach(producto => {
+                console.log("entra3");
+                if (producto.id == productoCarrito.productId) {
+                  let pathImagenDetalle: any[] = [];
+                  pathImagenDetalle = this.productoService.imagenesDetalle(
+                    producto.id
+                  );
+                  producto.idCarrito = productoCarrito.id;
+                  producto.path = pathImagenDetalle;
+                  producto.talle = productoCarrito.talle;
+                  this.productosCarrito.push(producto);
+                }
+              });
+            }
+          });
+        }
+      }, 2000);
     }
   }
 
@@ -76,9 +96,11 @@ export class SideCartComponent implements OnInit, DoCheck {
       if (localStorage.getItem('carritoDeCompras')) {
         productoCarrito = JSON.parse(localStorage.getItem('carritoDeCompras'));
 
+        console.log(productoCarrito);
+
         let total = 0;
         todosLosProductosJson.forEach(producto => {
-          if (producto.id == productoCarrito[0].productId) {
+          if (producto.id == productoCarrito[0].productId && productoCarrito[0].orden_id === 0) {
             this.productosCarrito = [];
             let pathImagenDetalle: any[] = [];
             pathImagenDetalle = this.productoService.imagenesDetalle(producto.id);
@@ -114,7 +136,7 @@ export class SideCartComponent implements OnInit, DoCheck {
             localStorage.getItem("todosLosProductos")
           );
           todosLosProductosJson.forEach(producto => {
-            if (producto.id == productoCarrito.productId) {
+            if (producto.id == productoCarrito.productId && productoCarrito.orden_id === 0) {
               let pathImagenDetalle: any[] = [];
               pathImagenDetalle = this.productoService.imagenesDetalle(producto.id);
               // this.productoService.getImagenesDetalle(producto.id).subscribe((res: any) => {
@@ -137,7 +159,7 @@ export class SideCartComponent implements OnInit, DoCheck {
     if (this.productosCarrito) {
       let total = 0;
       this.productosCarrito.forEach(productoCarrito => {
-        if (productoCarrito.descuento) {
+        if (productoCarrito.descuento && productoCarrito.orden_id === 0) {
           let descuento = ( (productoCarrito.descuento) / 100) * productoCarrito.precio;
           total = total + (productoCarrito.precio - descuento);
         } else {
@@ -150,12 +172,11 @@ export class SideCartComponent implements OnInit, DoCheck {
 
   quitarProducto(idCarrito) {
     this.productosCarrito.forEach((producto: any, index) => {
-      if (producto.idCarrito === idCarrito) {
+      if (producto.idCarrito === idCarrito && producto.orden_id === 0) {
         this.productosCarrito['total'] = this.productosCarrito['total'] - producto.precio;
         this.productosCarrito.splice(index, 1);
         this.carritoService.deleteProductoCarrito(idCarrito);
         this.carritoService.deleteCarritoBD(idCarrito).subscribe( res => {
-          console.log(res);
           return res;
         });
         if (this.carritoService.getCarrito() !== null) {
